+++ CHANGELOG.md
@@ -8,3 +8,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Added initial implementation
+
+### Fixed
+- Used `profile` query check whether user is logged in